api/requirements.txt
python-dotenv
flask
waitress
psycopg2-binary